Check image size before writing

+++ boot-files/boot/grub/i386-pc/rewrite-mbr.sh
@@ -1,8 +1,5 @@
 #!/bin/sh -e
 
-target=$1
-core_img=$2
-
 show_help()
 {
 	cat >&2 <<EOF
@@ -13,6 +10,9 @@ Arguments:
   core_image            the proper 'core.img.xxx' image
 EOF
 }
+
+target=$1
+core_img=$2
 
 # Check target
 if [ -f "$target" ]; then
@@ -34,15 +34,25 @@ if [ ! -f "$core_img" ]; then
 	exit 1
 fi
 
-set -x
 
 # MBR
 head -c446 boot.img > a
+
 # Partition table
-dd if=$target bs=512 count=1 of=b
+dd if=$target bs=512 count=1 of=b || exit 1
 tail -c66 b > c
+
 # Combine for the final image
 cat a c $core_img > d
-dd if=d of=$target bs=32k count=1 conv=notrunc
+
+# Ensure the size is <= 26K
+img_sz=`cat d | wc -c`
+if ! [ $img_sz -le 26624 ]; then
+	echo "*** Image size ($img_sz) exceeds 26K." >&2
+	exit 1
+fi
+
+# Write to target device or file
+dd if=d of=$target bs=1k count=26 conv=notrunc || exit 1
 
 rm -f a b c d
